// mobile/src/types.ts
export interface BedConfig {
  id: string
  name: string
  x: number
  z: number
  width: number
  depth: number
  color: string
  sunNeeds: 'full' | 'partial' | 'shade-tolerant'
  plantType?: PlantType
  rotation?: number
  hasTrellis?: boolean
  trellisHeight?: number
}

export interface BedAlert {
  bedId: string
  type: 'sun-exposure' | 'spacing'
  severity: 'warning' | 'error'
  message: string
}
// ...
export type PlantType =
  | 'tomatoes'
  | 'peppers'
  | 'cucumbers'
  | 'zucchini'
  | 'eggplant'
  | 'squash-winter'
  | 'beans'
  | 'peas'
  | 'bush-beans'
  | 'kale'
  | 'lettuce'
  | 'spinach'
  | 'swiss-chard'
  | 'arugula'
  | 'bok-choy'
  | 'broccoli'
  | 'cauliflower'
  | 'cabbage'
  | 'brussels-sprouts'
  | 'kohlrabi'
  | 'carrots'
  | 'beets'
  | 'radishes'
  | 'onions'
  | 'garlic'
  | 'potatoes'
  | 'herbs'
  | 'basil'
  | 'cilantro'
  | 'parsley'
// ...
export type MeasurementUnit = 'ft' | 'm'

export type StructureType = 'building' | 'fence'

export interface Structure {
  id: string
  type: StructureType
  name: string
  position: [number, number, number]
  size: [number, number, number]
  color: string
  castShadow: boolean
  receiveShadow: boolean
}

export interface GardenConfig {
  name: string
  savedAt: string
  season: string
  dayOfYear: number
  hour: number
  yardHeadingDeg: number
  structures: Structure[]
  beds: BedConfig[]
  showMeasurements: boolean
  measurementUnit: MeasurementUnit
  gardenItemsOpacity: number
  sunVizRadius?: number
  cityKey?: string
  latitude?: number
  longitude?: number
  cityTz?: string
  showGrid?: boolean
  gridSpacing?: number
  gridCenterX?: number
  gridCenterZ?: number
  gridWidth?: number
  gridDepth?: number
  heatmaps?: HeatmapInstance[]
  // Legacy single-heatmap fields (pre multi-heatmap; migrated on load)
  heatmapVisible?: boolean
  heatmapMode?: string
  heatmapOpacity?: number
  heatmapCenterX?: number
  heatmapCenterZ?: number
  heatmapWidth?: number
  heatmapDepth?: number
}